+++ buildchain/tailwind.config.js
@@ -2,7 +2,7 @@
 module.exports = {
   content: [
     '../src/templates/**/*.{twig,html}',
-    '../src/assetbundles/seomatic/src/**/*.{vue,html,js}',
+    './src/vue/**/*.{vue,html}',
   ],
   theme: {
   },