+++ src/controllers/SettingsController.php
@@ -117,7 +117,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Enabled sites
@@ -242,7 +242,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Basic variables
@@ -409,6 +409,7 @@ class SettingsController extends Controller
             }
 
             if ($hasErrors) {
+                Craft::error(print_r($metaBundle->metaGlobalVars->getErrors(), true), __METHOD__);
                 Craft::$app->getSession()->setError(Craft::t('app', "Couldn't save settings due to a Twig error."));
                 // Send the redirect back to the template
                 Craft::$app->getUrlManager()->setRouteParams([
@@ -452,7 +453,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Basic variables
@@ -534,7 +535,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Enabled sites
@@ -709,7 +710,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Basic variables
@@ -849,7 +850,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Basic variables
@@ -931,7 +932,7 @@ class SettingsController extends Controller
             Craft::error($e->getMessage(), __METHOD__);
         }
         $variables['baseAssetsUrl'] = Craft::$app->assetManager->getPublishedUrl(
-            '@nystudio107/seomatic/assetbundles/seomatic/dist',
+            '@nystudio107/seomatic/web/assets/dist',
             true
         );
         // Basic variables